Renamed spherical_intersect/reproject.py to spherical_intersect/core.py
