+++ reproject/__init__.py
@@ -12,4 +12,4 @@ from ._astropy_init import *
 if not _ASTROPY_SETUP_:
     from .high_level import reproject
     from .interpolation import reproject_interpolation
-    from .flux_conserving import reproject_flux_conserving
+    from .spherical_intersect import reproject_flux_conserving
