+++ reproject/spherical_intersect/core.py
